% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAllCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create the exposure and outcome cohorts}
\usage{
createCohorts(databaseDetails, outputFolder)
}
\arguments{
\item{databaseDetails}{The databaseDetails}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create the exposure and outcome cohorts
}
\details{
This function will create the exposure and outcome cohorts following the definitions included in
this package.
}
