% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{createStudyFiles}
\alias{createStudyFiles}
\title{Get the cohorts from atlas inserted into the package}
\usage{
createStudyFiles(baseUrl, packageDirectory, predictionAnalysisListFile)
}
\arguments{
\item{baseUrl}{The connection to the atlas webAPI e.g., http://api.ohdsi.org:80/WebAPI}

\item{packageDirectory}{(Default is working directory) The directory of the of the package containing the predictionAnalysisList.json when predictionAnalysisListFile is missing}

\item{predictionAnalysisListFile}{(optional) The json extracted from atlas.  If not
entered then it searches in inst/settings for predictionAnalysisList.json}
}
\description{
Get the cohorts from atlas inserted into the package
}
\details{
This function inserts the cohorts specified in the prediction analysis list
into the study package.
}
\examples{
\dontrun{
library(SkeletonPredictionStudy)
createStudyFiles(baseUrl='http://api.ohdsi.org:80/WebAPI')
}

}
