% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateShinyApp.R
\name{populateShinyApp}
\alias{populateShinyApp}
\title{Adds the results of the study into a shiny app}
\usage{
populateShinyApp(
  outputDirectory = "./ShinyApp",
  shinyDirectory,
  resultDirectory,
  minCellCount = 10,
  databaseName = "sharable name of development data"
)
}
\arguments{
\item{outputDirectory}{Name of folder containing the shiny app}

\item{shinyDirectory}{Location of a template shiny app (with no data)}

\item{resultDirectory}{The location of the results from the study}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included in the results.}

\item{databaseName}{A friendly name for the database}
}
\description{
Adds the results of the study into a shiny app
}
\details{
This function adds the results of the study into a shiny app ready for sharing
}
