% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubmitResults.R
\name{submitResults}
\alias{submitResults}
\title{Submit the study results to the study coordinating center}
\usage{
submitResults(outputFolder, privateKeyFileName, remoteFolder, userName)
}
\arguments{
\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{privateKeyFileName}{A character string denoting the path to an RSA private key.}

\item{remoteFolder}{The remote folder to upload the file to.}

\item{userName}{A character string containing the user name.}
}
\value{
TRUE if the upload was successful.
}
\description{
Submit the study results to the study coordinating center
}
\details{
This will upload the file \code{StudyResults.zip} to the OHDSI SFTP server
This requires an active internet connection.
}
